Assertions, whitespace, documentation
Assertions ensure that the bot is logged-in and/or flagged as bot while being logged-in. After this change, an option ``"assertions": "bot|user"`` is available for configuration.
Only adding assertions to requests changing content; one could consider enabling them for all requests, except login. Opinions?

We possibly want to make ``addAssertions`` private? Not sure, also open for discussion in this regard.

TAB-indention in ``.getBacklinks()``

JSDuck-documenting new methods.