build: Add Node 12 testing. Discontinue Node 8 and earlier
* The Node 4 LTS reached end-of-life in April 2018.
* The Node 6 LTS reached end-of-life in April 2019.
* The Node 8 LTS reached enf-of-life in December 2019.

<https://nodejs.org/en/about/releases/>
<https://github.com/nodejs/Release#end-of-life-releases>

This is done as preparation before other contributions I plan
to make to update dependencies (some of which no longer support
older versions), and some simplifications (some of which would
require additional packages be added to support older versions).

The previous versions of this package naturally conitinue to
be available for those using them on older Node.js versions,
and in the event of a critical bug we could even create an
"v0.x-release" branch or something and create a new patch release
from there against the older code.