bot: use arrow functions to preserve lexical 'this' instead of 'self'
None of the methods involved in this module make use of a custom
'this' value inside the callback, so let the same be consistently
available throughout a given method by using arrow functions.

Arrow functions, unlike regular functions, do not offer the function
its own "function" scope (with its own 'this' and 'arguments' vars),
and instead offer only a block scope.